requirements.txt
django
djangorestframework
pexels_api
djangorestframework_simplejwt
psycopg2
pyjwt
